+++ CHANGELOG.md
@@ -9,12 +9,39 @@ https://keepachangelog.com/en/1.0.0/
 https://www.markdownguide.org/basic-syntax/
 
 * [Unreleased](#unreleased)
+* [0.2.2](#022---2023-08-08)
 * [0.2.1](#021---2023-07-26)
 * [0.2.0](#020---2023-07-23)
 * [0.1.1](#011---2022-06-16)
 * [0.1.0](#010---2022-05-15)
 
 ## [Unreleased]
+
+## [0.2.2] - 2023-08-08
+
+### Added
+
+- Added DEFAULT_TENANT_ID constant in com.github.starnowski.posmulten.hibernate.common.Properties type [39](https://github.com/starnowski/posmulten-hibernate/issues/39)
+- Added com.github.starnowski.posmulten.hibernate.common.context.HibernateContext type [39](https://github.com/starnowski/posmulten-hibernate/issues/39)
+- Added com.github.starnowski.posmulten.hibernate.common.context.HibernateContextSupplier type [39](https://github.com/starnowski/posmulten-hibernate/issues/39)
+- Added com.github.starnowski.posmulten.hibernate.common.context.PropertiesHibernateContextFactory type [39](https://github.com/starnowski/posmulten-hibernate/issues/39)
+- Setting default tenant in method com.github.starnowski.posmulten.hibernate.hibernate5.connections.SharedSchemaMultiTenantConnectionProvider#getAnyConnection() [39](https://github.com/starnowski/posmulten-hibernate/issues/39)
+- Added com.github.starnowski.posmulten.hibernate.hibernate5.context.Hibernate5ContextSupplier type [39](https://github.com/starnowski/posmulten-hibernate/issues/39)
+- Added com.github.starnowski.posmulten.hibernate.hibernate5.context.Hibernate5ContextSupplierInitiator type [39](https://github.com/starnowski/posmulten-hibernate/issues/39)
+- Setting default tenant in method com.github.starnowski.posmulten.hibernate.hibernate6.connection.SharedSchemaMultiTenantConnectionProvider#getAnyConnection() [39](https://github.com/starnowski/posmulten-hibernate/issues/39)
+- Added com.github.starnowski.posmulten.hibernate.hibernate6.context.Hibernate6ContextSupplier type [39](https://github.com/starnowski/posmulten-hibernate/issues/39)
+- Added com.github.starnowski.posmulten.hibernate.hibernate6.context.Hibernate6ContextSupplierInitiator type [39](https://github.com/starnowski/posmulten-hibernate/issues/39)
+
+### Changed
+
+- Moved com.github.starnowski.posmulten.hibernate.hibernate5.Properties type to com.github.starnowski.posmulten.hibernate.common package [39](https://github.com/starnowski/posmulten-hibernate/issues/39)
+
+### Fixed
+
+- Closing connection by ConnectionProvider object in method com.github.starnowski.posmulten.hibernate.hibernate5.connections.SharedSchemaMultiTenantConnectionProvider#releaseAnyConnection(java.sql.Connection) [37](https://github.com/starnowski/posmulten-hibernate/issues/37)
+- Closing connection by ConnectionProvider object in method com.github.starnowski.posmulten.hibernate.hibernate5.connections.SharedSchemaMultiTenantConnectionProvider#releaseConnection(java.lang.String, java.sql.Connection) [37](https://github.com/starnowski/posmulten-hibernate/issues/37)
+- Closing connection by ConnectionProvider object in method com.github.starnowski.posmulten.hibernate.hibernate6.connection.SharedSchemaMultiTenantConnectionProvider#releaseAnyConnection(java.sql.Connection) [37](https://github.com/starnowski/posmulten-hibernate/issues/37)
+- Closing connection by ConnectionProvider object in method com.github.starnowski.posmulten.hibernate.hibernate6.connection.SharedSchemaMultiTenantConnectionProvider#releaseConnection(java.lang.String, java.sql.Connection) [37](https://github.com/starnowski/posmulten-hibernate/issues/37)
 
 ## [0.2.1] - 2023-07-26